% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr01_association_matrices.R
\name{association}
\alias{association}
\alias{arma}
\alias{matern}
\alias{cluster}
\title{Association Matrices for the BCNSM Marginal Regression Models}
\usage{
arma(p = 0L, q = 0L)

matern(D, smt = 0.5)

cluster(
  id,
  type = c("nonassociative", "ar1", "ma1", "uniform", "unstructured")
)
}
\arguments{
\item{p, q}{Order of the autoregressive and the moving average component,
respectively, passed as arguments to \code{arma()}.}

\item{D}{Matrix with values of the distances between pairs of data
locations for spatial data.}

\item{smt}{Value of the shape parameter of the Matérn correlation class.
The default \code{smt = 0.5} corresponds to an exponential correlation
model.}

\item{id}{Subject id for longitudinal/clustered data. This is a vector of
the same length of the number of observations. Please note that data
must be sorted in way that observations from the same cluster are
contiguous.}

\item{type}{A character string specifying the correlation structure among
groups for longitudinal/clustered data. At the
moment, the following are implemented:
\tabular{ll}{
\code{nonassociative}  \tab Non-associative. \cr
\code{ar1}  \tab Autoregressive of order 1. \cr
\code{ma1}  \tab Moving average of order 1. \cr
\code{uniform1}  \tab Uniform or exchangeable. \cr
\code{unstructured}  \tab Unstructured \cr
}}
}
\description{
Current available association matrices in the \code{bcnsm.reg}
    package for the class of the BCNSM marginal regression models.
}
\details{
The functions related to the association structures are a direct
    adaptation of the functions of the \code{\link[gcmr]{gcmr}} package.
    The documentation of the original functions of the package can be seen
    at \code{\link[gcmr]{cormat.gcmr}} documentation. The available
    association matrices are:
 \tabular{ll}{
 \bold{Function}  \tab \bold{Dependence structure}\cr
 \code{nonassociative}  \tab Non-associative structure. \cr
 \code{arma}  \tab ARMA(p, q). \cr
 \code{cluster}  \tab Longitudinal/clustered data. \cr
 \code{matern}  \tab Matérn spatial correlation. \cr
 }

 @return

 Each type of structure will require different arguments for the association
 matrix to be completed. However, all functions return a list of the
 following components:
 \itemize{
 \item{\code{npar}:}{ Number of parameters associated with the correlation
     structure.}
 \item{\code{start}:}{ A function \code{function(y)} that returns the
     initial value for use in optimization algorithms. Its argument is the
     vector/matrix of observations \code{y}.}
 \item{\code{Gamma}:}{ A function \code{function(alpha, d)} which returns the
     association matrix itself. Its arguments are the parameter vector
     associated with the correlation matrix structure and the dimension
     of the correlation matrix.}
 \item{\code{name}:}{ Name of the specified correlation structure.}
 }
}
\references{
Guido Masarotto, & Cristiano Varin (2017). Gaussian Copula
    Regression in R. Journal of Statistical Software, 77(8), 1--26.
}
\author{
Rodrigo M. R. de Medeiros <\email{rodrigo.matheus@live.com}>
}
