% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr04_methods.R
\name{marginreg-methods}
\alias{marginreg-methods}
\alias{model.frame.marginreg}
\alias{model.matrix.marginreg}
\alias{coef.marginreg}
\alias{vcov.marginreg}
\alias{logLik.marginreg}
\alias{AIC.marginreg}
\alias{print.marginreg}
\alias{summary.marginreg}
\alias{print.summary.marginreg}
\title{Methods for "marginreg" objects}
\usage{
\method{model.frame}{marginreg}(formula, ...)

\method{model.matrix}{marginreg}(object, model = c("mu", "sigma"), ...)

\method{coef}{marginreg}(object, model = c("mu", "sigma"), ...)

\method{vcov}{marginreg}(
  object,
  parameter = c("full", "mu", "sigma", "lambda", "nu", "gamma"),
  ...
)

\method{logLik}{marginreg}(object, ...)

\method{AIC}{marginreg}(object, ..., k = 2)

\method{print}{marginreg}(x, ...)

\method{summary}{marginreg}(object, ...)

\method{print}{summary.marginreg}(x, ...)
}
\arguments{
\item{formula}{a model formula or terms object or an R object.}

\item{...}{further arguments passed to or from other methods.}

\item{model}{a character indicating which submodel the function should act on. It can be
\code{"mu"} (default) or \code{"sigma"} to specify the regression setings associated
with the \code{mu} or \code{sigma}, respectively.}

\item{parameter}{character; specifies which submatrix of the asymptotic covariance matrix of the
maximum likelihood estimators should be returned. The options are \code{"full"} (default),
\code{"mu"}, \code{"sigma"}, \code{"lambda"}, \code{"nu"}, and \code{"gamma"}.}

\item{k}{numeric, the penalty per parameter to be used; the default
\code{k = 2} is the classical AIC.}

\item{x, object}{an object of class \code{marginreg}.}
}
\description{
Methods for "marginreg" objects
}
\author{
Rodrigo M. R. de Medeiros <\email{rodrigo.matheus@live.com}>
}
