% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv04_methods.R
\name{multireg-methods}
\alias{multireg-methods}
\alias{model.frame.multireg}
\alias{model.matrix.multireg}
\alias{vcov.multireg}
\alias{logLik.multireg}
\alias{AIC.multireg}
\alias{print.multireg}
\alias{summary.multireg}
\alias{print.summary.multireg}
\title{Methods for "multireg" objects}
\usage{
\method{model.frame}{multireg}(formula, ...)

\method{model.matrix}{multireg}(object, ...)

\method{vcov}{multireg}(
  object,
  parameter = c("full", "beta", "sigma", "lambda", "nu", "gamma"),
  ...
)

\method{logLik}{multireg}(object, ...)

\method{AIC}{multireg}(object, ..., k = 2)

\method{print}{multireg}(x, ...)

\method{summary}{multireg}(object, ...)

\method{print}{summary.multireg}(x, ...)
}
\arguments{
\item{formula}{a simbolic description of the model to be fitted, of type
\code{cbind(y1, y2, ..., yd) ~ x1 + x2 + ... + xb}.}

\item{...}{further arguments passed to or from other methods.}

\item{parameter}{character; specifies which submatrix of the asymptotic covariance matrix of the
maximum likelihood estimators should be returned. The options are \code{"full"} (default),
\code{"beta"}, \code{"sigma"}, \code{"lambda"}, \code{"nu"}, and \code{"gamma"}.}

\item{k}{numeric, the penalty per parameter to be used; the default
\code{k = 2} is the classical AIC.}

\item{x, object}{an object of class \code{"multireg"}.}
}
\description{
Methods for "multireg" objects
}
\author{
Rodrigo M. R. de Medeiros <\email{rodrigo.matheus@live.com}>
}
