% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr04_methods.R
\name{plot.marginreg}
\alias{plot.marginreg}
\title{Diagnostic plots for the BCNSM marginal regression models}
\usage{
\method{plot}{marginreg}(
  x,
  which = 1:4,
  type = c("conditional", "quantile"),
  ask = prod(graphics::par("mfcol")) < length(which) && grDevices::dev.interactive(),
  ...
)
}
\arguments{
\item{x}{an object of class \code{"marginreg"}.}

\item{which}{numeric. If a subset of the plots is required, specify a subset
of the numbers \\code{1:6}. See details below.}

\item{type}{character; it specifies which residual should be extracted.
The available arguments are \code{"conditional"} (default), \code{"quantile"}, and
\code{"epsilon"} (standardized epsilon's).}

\item{ask}{logical. If \code{TRUE}, the user is asked before each plot.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Diagnostic plots for the BCNSM marginal regression models
}
\details{
The \code{which} argument can be used to select a subset of currently seven supported
    types of displays. The displays are:

\describe{
  \item{\code{which = 1}:}{residuals vs fitted \code{mu} values.}
  \item{\code{which = 2}:}{residuals vs indices of observations.}
  \item{\code{which = 3}:}{residuals density vs standard normal density.}
  \item{\code{which = 4}:}{normal probability plot of the residuals.}
  \item{\code{which = 5}:}{autocorrelation plot of the residuals.}
  \item{\code{which = 6}:}{partial autocorrelation plot of the residuals.}
 }
}
