% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv04_methods.R
\name{plot.multireg}
\alias{plot.multireg}
\title{Visualization of the fit of the BCNSM distributions}
\usage{
\method{plot}{multireg}(
  x,
  type = c("quantile", "marginal", "mahalanobis", "epsilon"),
  levels = c(0.1, 0.01, 0.001, 1e-04),
  panel = NULL,
  ...
)
}
\arguments{
\item{x}{an object of class \code{multireg}.}

\item{type}{character; specifies which graphical should be produced. The available options
are \code{"quantile"} (default), \code{"marginal"}, \code{"mahalanobis"}, and
\code{"epsilon"}.}

\item{levels}{levels for contours plots. Used only when \code{type = "epsilon"}.}

\item{panel}{a vector of the form \code{c(nr, nc)} with the number of rows and columns, where
the figures will be drawn in an \code{nr-}by\code{-nc} array on the device. Used
only when \code{type = "marginal"}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Five plots (selectable by \code{type}) are currently available:
a plot of residuals against fitted values, a plot of residuals against
the indexes, a Normal Q-Q plot, a barplot with comparisons of the
observed and fitted frequencies, and a plot of the sample autocorelations
of the residuals.
}
\author{
Rodrigo M. R. de Medeiros <\email{rodrigo.matheus@live.com}>
}
