% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr04_methods.R
\name{predict.marginreg}
\alias{predict.marginreg}
\title{Predict Method for BCNSM Marginal Regression Fits}
\usage{
\method{predict}{marginreg}(
  object,
  newdata = NULL,
  at = 0.5,
  na.action = stats::na.pass,
  h,
  dist,
  ...
)
}
\arguments{
\item{object}{an \code{"marginreg"} object.}

\item{newdata}{optionally, a data frame in which to look for variables
with which to predict. If omitted, the fitted quantiles are returned.}

\item{at}{the order of the quantile to be predicted. The default is to predict the median,
that is, \code{at = 0.5}.}

\item{na.action}{function determining what should be done with missing
values in \code{newdata}. The default is to predict \code{NA}.}

\item{h}{number of steps forward for time series forecasting.}

\item{dist}{a distance matrix used specifically for predictions of spatial data at locations not
originally observed in the data set. Each row of the matrix is assumed to contain the distances
from the original observations to an unobserved location.}

\item{...}{arguments passed to or from other methods.}
}
\value{
A vector of predictions.
}
\description{
Obtains predictions from a fitted BCNSM marginal regression object.
}
