% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr04_methods.R
\name{residuals.marginreg}
\alias{residuals.marginreg}
\title{Extract Model Residuals}
\usage{
\method{residuals}{marginreg}(object, type = c("conditional", "quantile"), ...)
}
\arguments{
\item{object}{a \code{"marginreg"} object.}

\item{type}{character; it specifies which residual should be extracted.
The available arguments are \code{"conditional"} (default) and \code{"quantile"}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Extract Model Residuals
}
