% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv04_methods.R
\name{residuals.multireg}
\alias{residuals.multireg}
\title{Extract Model Residuals for a Fitted BCNSM Regression}
\usage{
\method{residuals}{multireg}(object, type = c("quantile", "marginal", "epsilon"), ...)
}
\arguments{
\item{object}{an \code{"multireg"} object.}

\item{type}{character; specifies which residual should be extracted. The available arguments are
\code{"quantile"} (default), \code{"mahalanobis"}, and \code{"epsilon"}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Extract Model Residuals for a Fitted BCNSM Regression
}
\author{
Rodrigo M. R. de Medeiros <\email{rodrigo.matheus@live.com}>
}
